// src/pages/types/TVShow.types.ts
// Generated by https://quicktype.io

export interface TVMazeShow {
  id: number;
  url: string;
  name: string;
  type: string;
  language: string;
  genres: string[];
  status: string;
  runtime: number;
  premiered: string;
  officialSite: string;
  schedule: Schedule;
  rating: Rating;
  weight: number;
  network: Network;
  webChannel: null;
  externals: Externals;
  image: Image;
  summary: string;
  updated: number;
  _links: Links;
}

interface Links {
  self: Previousepisode;
  previousepisode: Previousepisode;
}

interface Previousepisode {
  href: string;
}

interface Externals {
  tvrage: null;
  thetvdb: number;
  imdb: string;
}

interface Image {
  medium: string;
  original: string;
}

interface Network {
  id: number;
  name: string;
  country: Country;
}

interface Country {
  name: string;
  code: string;
  timezone: string;
}

interface Rating {
  average: number;
}

interface Schedule {
  time: string;
  days: string[];
}